+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Recommendation: for ease of reading, use the following order:
 - Fixed
 -->
 
+## [Unreleased]
+### Changed
+- `kamu repo alias list`: added JSON output alongside with other formats mentioned in the command's help
+
 ## [0.204.4] - 2024-09-30
 ### Changed
 - CLI command tweaks: